requirements.txt
requests
beautifulsoup4
feedparser
google-generativeai
slack_sdk
python-dotenv
schedule 
google.generativeai